Fix for the problem
Some experiments later I figured out how to get around the issue. The
problem is the `&mut self` part of the function that is exposed to JS,
we needed it because the function had to modify an internal variable.

Wrapping the variable in an Arc<Mutex<>> allows us to remove the `mut`
and change the function parameters back to just `&self` and the error
magically goes away.